feat(arch): migrate kernel from i386 to x86_64
- Rewrite kernel entry (head.S) to set up 4-level page tables (PML4),
  enable PAE/long mode, and jump to 64-bit code with higher-half
  mapping at 0xFFFFFFFF80000000
- Update TrapFrame and TrapRegisters to use 64-bit registers (r8-r15,
  rax-rdi) with 16-byte IDT entries and iretq
- Convert context switch (switch.S) to x86_64 calling convention
  using callee-saved registers (rbx, rbp, r12-r15) via rdi/rsi args
- Implement 4-level page table walk (PML4→PDPT→PD→PT) in get_pte()
- Widen ELF structures, MMIO base types, and size_t for 64-bit;
  replace x86 inline asm do_div with portable C implementation
- Update kernel.ld to ELF64 output at KERNEL_BASE 0xFFFFFFFF80000000
- Switch QEMU target to qemu-system-x86_64 with 128M memory

+++ kern/arch/x86/e820.h
@@ -16,7 +16,7 @@ struct e820map {
 
 template<typename F>
 void traverse_e820_map(F&& callback) {
-    e820map *map = reinterpret_cast<e820map*>(E820_MEM_BASE + KERNEL_BASE);
+    e820map *map = reinterpret_cast<e820map*>((uintptr_t)E820_MEM_BASE + KERNEL_BASE);
     for (int i = 0; i < map->nr_map; i++) {
         callback(map->map[i].addr, map->map[i].size, map->map[i].type);
     }
